requirements.txt
requests>=2.31.0
pyyaml>=6.0
gtts>=2.5.0
openai>=1.0.0

